CRT540: move some function out of core
