+++ package.json
@@ -11,8 +11,8 @@
         "build": "npm run build -w crt",
         "build:example": "npm run build -w example",
         "nuke": "node scripts/nuke.js",
-    "test": "npm run build && vitest run",
-    "test:watch": "npm run build && vitest watch",
+        "test": "npm run build && vitest run",
+        "test:watch": "npm run build && vitest watch",
         "coverage": "node scripts/coverage.js",
         "lint:js": "eslint .",
         "lint:prettier": "prettier --write .",
