+++ packages/crt/src/index.js
@@ -23,6 +23,7 @@ export { $dataGet } from './utils/dom/$dataGet.js';
 export { $dataSet } from './utils/dom/$dataSet.js';
 export { normaliseEventTarget } from './utils/dom/normaliseEventTarget.js';
 export { removeElement } from './utils/dom/removeElement.js';
+export { cx } from './utils/dom/cx.js';
 
 // Utils - Input
 export { assertKey, getDirectionFromKeyCode } from './utils/keys.js';
@@ -33,19 +34,14 @@ export { getBaseFontSize } from './utils/dom/getBaseFontSize.js';
 export { pxToRem } from './utils/style/pxToRem.js';
 export { transformProp } from './utils/style/prefix.js';
 
-// Utils - Number
-export { parseDecimal } from './utils/math/parseDecimal.js';
-
 export { collectionToArray } from './utils/dom/collectionToArray.js';
-export { cx } from './utils/dom/cx.js';
-export { toTitleCase } from './utils/string/toTitleCase.js';
-export { parseSearchParams } from './utils/dom/parseSearchParams.js';
 
 export { throttle } from './utils/function/throttle.js';
+export { debounce } from './utils/function/debounce.js';
 export { noop } from './utils/function/noop.js';
 
 // VDOM
 export { diff, stringToHTML } from './differenceEngine.js';
 
-// Scroll
-export { clearDeadSeaCache, scrollAction } from './deadSea.js';
+// async
+export { request } from './utils/async/request.js';
