+++ packages/example/src/libs/router/utils/getLocationHashValue.js
