+++ packages/example/src/views/search.js
@@ -37,10 +37,9 @@ function handleKeyboard(event) {
     const elTarget = normaliseEventTarget(event);
     if (
         elTarget instanceof HTMLElement &&
-        (
-            event instanceof MouseEvent ||
-            event instanceof KeyboardEvent && assertKey(event, AdditionalKeys.ENTER)
-        )
+        (event instanceof MouseEvent ||
+            (event instanceof KeyboardEvent &&
+                assertKey(event, AdditionalKeys.ENTER)))
     ) {
         const keyPressValue = $dataGet(elTarget, 'keyValue');
 
@@ -118,10 +117,7 @@ function updateSearchList() {
                             href: res.url,
                             id: res.id,
                         },
-                        div(
-                            { className: s.searchResult },
-                            span({}, res.title)
-                        )
+                        div({ className: s.searchResult }, span({}, res.title))
                     )
                 )
             );
@@ -200,11 +196,7 @@ export function createSearchView(options) {
             return div(
                 { className: 'view', id: this.id },
                 div({ className: s.searchInput, id: 'search-input' }),
-                div(
-                    { className: s.panels2 },
-                    this.keyboard,
-                    this.searchResults
-                )
+                div({ className: s.panels2 }, this.keyboard, this.searchResults)
             );
         },
     };
